phase3b_requirements.txt
polygon-api-client
requests
beautifulsoup4
lxml
GoogleNews
